+++ lib/cisco_node_utils/bgp_neighbor_af.rb
@@ -756,5 +756,25 @@ module Cisco
     def default_weight
       config_get_default('bgp_neighbor_af', 'weight')
     end
+
+    def rewrite_evpn_rt_asn
+      config_get('bgp_neighbor_af', 'rewrite_evpn_rt_asn', @get_args)
+    end
+
+    def rewrite_evpn_rt_asn=(state)
+      Feature.nv_overlay_evpn_enable if state
+      set_args_keys(state: (state ? '' : 'no'))
+      if @set_args[:state] == 'no'
+        unless rewrite_evpn_rt_asn == default_rewrite_evpn_rt_asn
+          config_set('bgp_neighbor_af', 'rewrite_evpn_rt_asn', @set_args)
+        end
+      else
+        config_set('bgp_neighbor_af', 'rewrite_evpn_rt_asn', @set_args)
+      end
+    end
+
+    def default_rewrite_evpn_rt_asn
+      config_get_default('bgp_neighbor_af', 'rewrite_evpn_rt_asn')
+    end
   end
 end