+++ lib/cisco_node_utils/vxlan_vtep.rb
@@ -173,5 +173,29 @@ module Cisco
     def default_shutdown
       config_get_default('vxlan_vtep', 'shutdown')
     end
+
+    def multisite_border_gateway_interface
+      config_get('vxlan_vtep', 'multisite_bg_intf', name: @name)
+    end
+
+    def multisite_border_gateway_interface=(val)
+      set_args = { name: @name }
+      set_args[:state] = val.empty? ? 'no' : ''
+      set_args[:lpbk_intf] = val.empty? ? 'loopback0' : val
+      if set_args[:state] == 'no'
+        # 'no multisite border-gateway' doesn't work without interface
+        # defaulting to 'loopback0' still clears any configuration
+        intf = multisite_border_gateway_interface
+        unless intf == default_multisite_border_gateway_interface
+          config_set('vxlan_vtep', 'multisite_bg_intf', set_args)
+        end
+      else
+        config_set('vxlan_vtep', 'multisite_bg_intf', set_args)
+      end
+    end
+
+    def default_multisite_border_gateway_interface
+      config_get_default('vxlan_vtep', 'multisite_bg_intf')
+    end
   end  # Class
 end    # Module